Refactor logical table metadata cache.
