+++ pinot-server/src/main/java/org/apache/pinot/server/starter/helix/HelixInstanceDataManager.java
@@ -39,21 +39,14 @@ import java.util.concurrent.locks.Lock;
 import java.util.function.Supplier;
 import javax.annotation.Nullable;
 import javax.annotation.concurrent.ThreadSafe;
-import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang3.tuple.Pair;
-import org.apache.helix.AccessOption;
 import org.apache.helix.HelixManager;
 import org.apache.helix.store.zk.ZkHelixPropertyStore;
 import org.apache.helix.zookeeper.datamodel.ZNRecord;
-import org.apache.helix.zookeeper.zkclient.IZkChildListener;
-import org.apache.helix.zookeeper.zkclient.IZkDataListener;
 import org.apache.pinot.common.metadata.ZKMetadataProvider;
 import org.apache.pinot.common.metrics.ServerMetrics;
 import org.apache.pinot.common.restlet.resources.SegmentErrorInfo;
-import org.apache.pinot.common.utils.LogicalTableConfigUtils;
-import org.apache.pinot.common.utils.SchemaUtils;
-import org.apache.pinot.common.utils.config.TableConfigUtils;
 import org.apache.pinot.core.data.manager.InstanceDataManager;
 import org.apache.pinot.core.data.manager.LogicalTableContext;
 import org.apache.pinot.core.data.manager.provider.TableDataManagerProvider;
@@ -77,7 +70,6 @@ import org.apache.pinot.spi.data.LogicalTableConfig;
 import org.apache.pinot.spi.data.Schema;
 import org.apache.pinot.spi.env.PinotConfiguration;
 import org.apache.pinot.spi.plugin.PluginManager;
-import org.apache.pinot.spi.utils.CommonConstants.ZkPaths;
 import org.apache.pinot.spi.utils.TimestampIndexUtils;
 import org.apache.pinot.spi.utils.builder.TableNameBuilder;
 import org.apache.zookeeper.data.Stat;
@@ -94,18 +86,8 @@ public class HelixInstanceDataManager implements InstanceDataManager {
 
   private final Map<String, TableDataManager> _tableDataManagerMap = new ConcurrentHashMap<>();
 
-  // A cache for the logical table context
-  private static final String TABLE_CONFIG_PATH_PREFIX = "/CONFIGS/TABLE/";
-  private static final String SCHEMA_PATH_PREFIX = "/SCHEMAS/";
-
-  private final Map<String, LogicalTableConfig> _logicalTableConfigMap = new ConcurrentHashMap<>();
-  private final Map<String, Schema> _schemaMap = new ConcurrentHashMap<>();
-  private final Map<String, TableConfig> _tableConfigMap = new ConcurrentHashMap<>();
-  private final Map<String, List<Pair<String, String>>> _tableNameToLogicalTableNamesMap = new ConcurrentHashMap<>();
-
-  ZkTableConfigChangeListener _zkTableConfigChangeListener;
-  ZkSchemaChangeListener _zkSchemaChangeListener;
-  ZkLogicalTableConfigChangeListener _zkLogicalTableConfigChangeListener;
+  // Logical table metadata cache to cache logical table configs, schemas, and offline/realtime table configs.
+  private final LogicalTableMetadataCache _logicalTableMetadataCache = new LogicalTableMetadataCache();
 
   // TODO: Consider making segment locks per table instead of per instance
   private final SegmentLocks _segmentLocks = new SegmentLocks();
@@ -250,13 +232,8 @@ public class HelixInstanceDataManager implements InstanceDataManager {
   @Override
   public synchronized void start() {
     _propertyStore = _helixManager.getHelixPropertyStore();
-
-    _zkTableConfigChangeListener = new ZkTableConfigChangeListener();
-    _zkSchemaChangeListener = new ZkSchemaChangeListener();
-    _zkLogicalTableConfigChangeListener = new ZkLogicalTableConfigChangeListener();
-
-    // Add child listeners to the property store for logical table config changes
-    _propertyStore.subscribeChildChanges(ZkPaths.LOGICAL_TABLE_PARENT_PATH, _zkLogicalTableConfigChangeListener);
+    // Initialize logical table metadata cache
+    _logicalTableMetadataCache.init(_propertyStore);
 
     LOGGER.info("Helix instance data manager started");
   }
@@ -285,7 +262,8 @@ public class HelixInstanceDataManager implements InstanceDataManager {
       }
     }
     SegmentBuildTimeLeaseExtender.shutdownExecutor();
-    _propertyStore.unsubscribeChildChanges(ZkPaths.LOGICAL_TABLE_PARENT_PATH, _zkLogicalTableConfigChangeListener);
+    // shutdown logical table metadata cache
+    _logicalTableMetadataCache.shutdown();
     LOGGER.info("Helix instance data manager shut down");
   }
 
@@ -568,12 +546,12 @@ public class HelixInstanceDataManager implements InstanceDataManager {
   @Nullable
   @Override
   public LogicalTableContext getLogicalTableContext(String logicalTableName) {
-    Schema schema = _schemaMap.get(logicalTableName);
+    Schema schema = _logicalTableMetadataCache.getSchema(logicalTableName);
     if (schema == null) {
       LOGGER.warn("Failed to find schema for logical table: {}, skipping", logicalTableName);
       return null;
     }
-    LogicalTableConfig logicalTableConfig = _logicalTableConfigMap.get(logicalTableName);
+    LogicalTableConfig logicalTableConfig = _logicalTableMetadataCache.getLogicalTableConfig(logicalTableName);
     if (logicalTableConfig == null) {
       LOGGER.warn("Failed to find logical table config for logical table: {}, skipping", logicalTableName);
       return null;
@@ -581,7 +559,7 @@ public class HelixInstanceDataManager implements InstanceDataManager {
 
     TableConfig offlineTableConfig = null;
     if (logicalTableConfig.getRefOfflineTableName() != null) {
-      offlineTableConfig = _tableConfigMap.get(logicalTableConfig.getRefOfflineTableName());
+      offlineTableConfig = _logicalTableMetadataCache.getTableConfig(logicalTableConfig.getRefOfflineTableName());
       if (offlineTableConfig == null) {
         LOGGER.warn("Failed to find offline table config for logical table: {}, skipping", logicalTableName);
         return null;
@@ -590,208 +568,12 @@ public class HelixInstanceDataManager implements InstanceDataManager {
 
     TableConfig realtimeTableConfig = null;
     if (logicalTableConfig.getRefRealtimeTableName() != null) {
-      realtimeTableConfig = _tableConfigMap.get(logicalTableConfig.getRefRealtimeTableName());
+      realtimeTableConfig = _logicalTableMetadataCache.getTableConfig(logicalTableConfig.getRefRealtimeTableName());
       if (realtimeTableConfig == null) {
         LOGGER.warn("Failed to find realtime table config for logical table: {}, skipping", logicalTableName);
         return null;
       }
     }
     return new LogicalTableContext(logicalTableConfig, schema, offlineTableConfig, realtimeTableConfig);
-  }
-
-  private class ZkTableConfigChangeListener implements IZkDataListener {
-
-    @Override
-    public synchronized void handleDataChange(String path, Object data) {
-      if (data != null) {
-        ZNRecord znRecord = (ZNRecord) data;
-        try {
-          TableConfig tableConfig = TableConfigUtils.fromZNRecord(znRecord);
-          _tableConfigMap.put(tableConfig.getTableName(), tableConfig);
-        } catch (Exception e) {
-          LOGGER.error("Caught exception while refreshing table config for ZNRecord: {}", znRecord.getId(), e);
-        }
-      }
-    }
-
-    @Override
-    public synchronized void handleDataDeleted(String path) {
-      // no-op, table config should not be deleted while referenced in the logical table config
-    }
-  }
-
-  private class ZkSchemaChangeListener implements IZkDataListener {
-
-    @Override
-    public synchronized void handleDataChange(String path, Object data) {
-      if (data != null) {
-        ZNRecord znRecord = (ZNRecord) data;
-        try {
-          Schema schema = SchemaUtils.fromZNRecord(znRecord);
-          _schemaMap.put(schema.getSchemaName(), schema);
-        } catch (Exception e) {
-          LOGGER.error("Caught exception while refreshing schema for ZNRecord: {}", znRecord.getId(), e);
-        }
-      }
-    }
-
-    @Override
-    public synchronized void handleDataDeleted(String path) {
-      // no-op, schema should not be deleted before the logical table config
-    }
-  }
-
-  private class ZkLogicalTableConfigChangeListener implements IZkChildListener, IZkDataListener {
-
-    @Override
-    public synchronized void handleChildChange(String path, List<String> logicalTableNames) {
-      if (CollectionUtils.isEmpty(logicalTableNames)) {
-        return;
-      }
-
-      // Only process new added logical tables. Changed/removed logical tables are handled by other callbacks.
-      List<String> pathsToAdd = new ArrayList<>();
-      for (String logicalTableName : logicalTableNames) {
-        if (!_logicalTableConfigMap.containsKey(logicalTableName)) {
-          pathsToAdd.add(ZkPaths.LOGICAL_TABLE_PATH_PREFIX + logicalTableName);
-        }
-      }
-      if (!pathsToAdd.isEmpty()) {
-        addLogicalTableConfigs(pathsToAdd);
-      }
-    }
-
-    @Override
-    public synchronized void handleDataChange(String path, Object data) {
-      if (data != null) {
-        updateLogicalTableConfig((ZNRecord) data);
-      }
-    }
-
-    @Override
-    public synchronized void handleDataDeleted(String path) {
-      // NOTE: The path here is the absolute ZK path instead of the relative path to the property store.
-      String logicalTableName = path.substring(path.lastIndexOf('/') + 1);
-      removeLogicalTableConfig(logicalTableName);
-    }
-
-    private synchronized void addLogicalTableConfigs(List<String> pathsToAdd) {
-      for (String path : pathsToAdd) {
-        ZNRecord znRecord = _propertyStore.get(path, null, AccessOption.PERSISTENT);
-        if (znRecord != null) {
-          try {
-            LogicalTableConfig logicalTableConfig = LogicalTableConfigUtils.fromZNRecord(znRecord);
-            String logicalTableName = logicalTableConfig.getTableName();
-            _logicalTableConfigMap.put(logicalTableName, logicalTableConfig);
-
-            if (logicalTableConfig.getRefOfflineTableName() != null) {
-              addTableConfig(logicalTableConfig.getRefOfflineTableName(), logicalTableName);
-            }
-            if (logicalTableConfig.getRefRealtimeTableName() != null) {
-              addTableConfig(logicalTableConfig.getRefRealtimeTableName(), logicalTableName);
-            }
-
-            addSchema(logicalTableName);
-            String logicalTableConfigPath = ZkPaths.LOGICAL_TABLE_PATH_PREFIX + logicalTableName;
-            getPropertyStore().subscribeDataChanges(logicalTableConfigPath, _zkLogicalTableConfigChangeListener);
-          } catch (Exception e) {
-            LOGGER.error("Caught exception while refreshing logical table config for ZNRecord: {}", znRecord.getId(),
-                e);
-          }
-        }
-      }
-    }
-
-    private synchronized void addTableConfig(String tableName, String logicalTableName) {
-      TableConfig tableConfig = ZKMetadataProvider.getTableConfig(_propertyStore, tableName);
-      Preconditions.checkArgument(tableConfig != null, "Failed to find table config for table: %s", tableName);
-      _tableNameToLogicalTableNamesMap.computeIfAbsent(tableName, k -> new ArrayList<>())
-          .add(Pair.of(tableName, logicalTableName));
-      _tableConfigMap.put(tableName, tableConfig);
-      String path = TABLE_CONFIG_PATH_PREFIX + tableName;
-      getPropertyStore().subscribeDataChanges(path, _zkTableConfigChangeListener);
-    }
-
-    private synchronized void addSchema(String logicalTableName) {
-      Schema schema = ZKMetadataProvider.getSchema(getPropertyStore(), logicalTableName);
-      Preconditions.checkArgument(schema != null,
-          "Failed to find schema for logical table: %s", logicalTableName);
-      _schemaMap.put(schema.getSchemaName(), schema);
-      String schemaPath = SCHEMA_PATH_PREFIX + schema.getSchemaName();
-      getPropertyStore().subscribeDataChanges(schemaPath, _zkSchemaChangeListener);
-    }
-
-    private synchronized void updateLogicalTableConfig(ZNRecord znRecord) {
-      try {
-        LogicalTableConfig logicalTableConfig = LogicalTableConfigUtils.fromZNRecord(znRecord);
-        String logicalTableName = logicalTableConfig.getTableName();
-        LogicalTableConfig oldLogicalTableConfig = _logicalTableConfigMap.put(logicalTableName, logicalTableConfig);
-        Preconditions.checkArgument(oldLogicalTableConfig != null,
-            "Logical table config for logical table: %s should have been created before", logicalTableName);
-
-        // Remove the old table configs from the table config map
-        if (oldLogicalTableConfig.getRefOfflineTableName() != null
-            && !oldLogicalTableConfig.getRefOfflineTableName().equals(logicalTableConfig.getRefOfflineTableName())) {
-          removeTableConfig(oldLogicalTableConfig.getRefOfflineTableName(), logicalTableName);
-        }
-        if (oldLogicalTableConfig.getRefRealtimeTableName() != null
-            && !oldLogicalTableConfig.getRefRealtimeTableName().equals(logicalTableConfig.getRefRealtimeTableName())) {
-          removeTableConfig(oldLogicalTableConfig.getRefRealtimeTableName(), logicalTableName);
-        }
-
-        // Add the new table configs to the table config map
-        if (logicalTableConfig.getRefOfflineTableName() != null
-            && !logicalTableConfig.getRefOfflineTableName().equals(oldLogicalTableConfig.getRefOfflineTableName())) {
-          addTableConfig(logicalTableConfig.getRefOfflineTableName(), logicalTableName);
-        }
-        if (logicalTableConfig.getRefRealtimeTableName() != null
-            && !logicalTableConfig.getRefRealtimeTableName().equals(oldLogicalTableConfig.getRefRealtimeTableName())) {
-          addTableConfig(logicalTableConfig.getRefRealtimeTableName(), logicalTableName);
-        }
-      } catch (Exception e) {
-        LOGGER.error("Caught exception while refreshing logical table for ZNRecord: {}", znRecord.getId(), e);
-      }
-    }
-
-    private synchronized void removeLogicalTableConfig(String logicalTableName) {
-      LogicalTableConfig logicalTableConfig = _logicalTableConfigMap.remove(logicalTableName);
-      if (logicalTableConfig != null) {
-        // Remove the table configs from the table config map
-        String offlineTableName = logicalTableConfig.getRefOfflineTableName();
-        String realtimeTableName = logicalTableConfig.getRefRealtimeTableName();
-        if (offlineTableName != null) {
-          removeTableConfig(offlineTableName, offlineTableName);
-        }
-        if (realtimeTableName != null) {
-          removeTableConfig(realtimeTableName, logicalTableName);
-        }
-        // remove schema
-        removeSchema(logicalTableConfig);
-        // Unsubscribe from the logical table config path
-        String logicalTableConfigPath = ZkPaths.LOGICAL_TABLE_PATH_PREFIX + logicalTableName;
-        getPropertyStore().unsubscribeDataChanges(logicalTableConfigPath, _zkLogicalTableConfigChangeListener);
-      }
-    }
-
-    private synchronized void removeTableConfig(String tableName, String logicalTableName) {
-      _tableNameToLogicalTableNamesMap.computeIfPresent(tableName, (k, v) -> {
-        v.remove(Pair.of(tableName, logicalTableName));
-        return v;
-      });
-
-      if (_tableNameToLogicalTableNamesMap.getOrDefault(tableName, List.of()).isEmpty()) {
-        _tableNameToLogicalTableNamesMap.remove(tableName);
-        _tableConfigMap.remove(tableName);
-        String path = TABLE_CONFIG_PATH_PREFIX + tableName;
-        getPropertyStore().unsubscribeDataChanges(path, _zkTableConfigChangeListener);
-      }
-    }
-
-    private synchronized void removeSchema(LogicalTableConfig logicalTableConfig) {
-      String schemaName = logicalTableConfig.getTableName();
-      _schemaMap.remove(schemaName);
-      String schemaPath = SCHEMA_PATH_PREFIX + schemaName;
-      getPropertyStore().unsubscribeDataChanges(schemaPath, _zkSchemaChangeListener);
-    }
   }
 }
